refactor: ワープをリファクタリング
