+++ Group-F-UnityProject/Assets/Sprites/.gitignore
@@ -1,0 +1,1 @@
+!*.png
