+++ Group-F-UnityProject/Assets/_Completed-Assets/Scripts/Warp.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a252a9dc19fbefe4d9f41b591ac4c6f7
-PrefabImporter:
+guid: 32a9c672434b41948843c7a35eab6cd0
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
