--- Group-F-UnityProject/Assets/warp.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 25feaccadfabc994894f0dd31b0f460d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
