--- Group-F-UnityProject/Assets/warp/warp tp.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64cb06e56e03f854b9fee72c75c02978
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
