--- Group-F-UnityProject/Assets/warp/warp2.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 005a26e28c172f645a93ed855b527a5b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
